Int128 and UInt128

+++ src/compiler/crystal/repl/op_code.cr
@@ -2,7 +2,7 @@ require "./repl"
 require "./instructions"
 
 {% begin %}
-  enum Crystal::Repl::OpCode : UInt8
+  enum Crystal::Repl::OpCode : UInt16
     {% for name, instruction, i in Crystal::Repl::Instructions %}
       {{ name.id.upcase }} = {{ i }}
     {% end %}
